requirements.txt
sentry_sdk
bottle
gunicorn
python-dotenv
